Chat UI toolkit (#24)
* Created basic chat ui using ui toolkit

* what I did in class added asset NPC

* Got player working, now no chat overlay

* no

* Added sprites and cleaner UI

Last push was moving the text bubble. This push is adding sprites to NPC that aren't pre existing

* Nothing new

Moved the NPC and his color. Got it so only displays the chat when in range of NPC. However there is a duplicate bubble that is always there going to fix that tonight

* no more ghosts

Got rid of the duplicate. With NPC's will need to make new code maybe for each indivitual NPC. Or until I finagle more.

* NPC have their own script.

need to find NPC google docs to add the script

* Got the first NPC chat lines implimented

I would like help figuring out how to do check points for auto saying text

* pushing

* merger branch main

* npc messages back